backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
PyJWT>=2.8
python-dotenv>=1.0
